remove an unused getCapabilities function, move the validation of a missing block_name up front to where all the other validations are
