+++ config/test-fixtures/config.invalid12.yml
@@ -1,0 +1,20 @@
+version: "0.1"
+apps:
+  acme-docs:
+    module: nullstone/aws-fargate-service
+    vars:
+      num_tasks: 2
+    connections:
+      cluster-namespace:
+        block_name: namespace0
+    environment:
+      TESTING: abc123
+      BLAH: blahblahblah
+    capabilities:
+      - module: nullstone/aws-load-balancer
+        vars:
+          health_check_path: "/status"
+        connections:
+          subdomain:
+            stack_id: 1
+        namespace: primary
